backend/requirements.txt
# Core backend
flask
flask-cors
flask-limiter
python-dotenv

# Database
supabase

# RAG & AI
llama-index
llama-index-vector-stores-supabase
llama-index-readers-file
openai

# PDF extraction
llama-parse

# Web scraping
firecrawl-py

# Social media scraping
apify-client

# Admin password hashing
bcrypt

# Testing
pytest